R3 AR-1: extract parse detection and SHOW-clause parsing into submodules
Begin splitting the parse god-module (5.7k lines) into cohesive units.
Convert src/parse.rs into a directory module and move two self-contained
clusters out:

- parse::detect — keyword-prefix matching, comment/whitespace skipping,
  the longest-first DDL prefix table, the public detect_ddl_kind /
  detect_semantic_view_ddl entry points, and fuzzy near-miss suggestion.
- parse::show_clauses — LIKE / IN / FOR METRIC / STARTS WITH / LIMIT
  filter parsing for SHOW SEMANTIC statements plus WHERE/LIMIT suffix
  rendering.

Both are re-exported from parse/mod.rs (pub for the public API surface,
pub(crate) for internal helpers) so existing crate::parse::* call sites
and the in-module test suite resolve unchanged. extract_quoted_string
stays in mod.rs since it is shared across detection, rewrite, and
validation paths; show_clauses references it via super::.

No behaviour change. 1045 lib tests, clippy -D warnings, and fmt all
pass.
